// src/schemas/companySchemas.ts
import { z } from 'zod';

export const DatabaseCompanySchema = z.object({
    id: z.number().int().positive(),
    public_id: z.string().uuid(),
    trade_name: z.string(),
    company_name: z.string().nullable().optional(),
    cnpj: z.string().nullable().optional(),
    tax_regime: z.string().nullable().optional(),
    email: z.string().nullable().optional(),
    phone: z.string().nullable().optional(),
    zipcode: z.string().nullable().optional(),
    street: z.string().nullable().optional(),
    number: z.string().nullable().optional(),
    complement: z.string().nullable().optional(),
    neighborhood: z.string().nullable().optional(),
    city: z.string().nullable().optional(),
    state: z.string().nullable().optional(),
    api_token: z.string().nullable().optional(),
    swagger_api_token: z.string().nullable().optional(),
    certificate_name: z.string().nullable().optional(),
    certificate_url: z.string().nullable().optional(),
    certificate_expiration: z.union([z.string(), z.date()]).nullable().optional(),
    logo_url: z.string().nullable().optional(),
    logo_filename: z.string().nullable().optional(),
    logo_base64: z.string().nullable().optional(),
    serv_solidcon: z.string().nullable().optional(),
    bd_solidcon: z.string().nullable().optional(),
    login_solidcon: z.string().nullable().optional(),
    senha_solidcon: z.string().nullable().optional(),
    serv_alterdata: z.string().nullable().optional(),
    bd_alterdata: z.string().nullable().optional(),
    login_alterdata: z.string().nullable().optional(),
    senha_alterdata: z.string().nullable().optional(),
    porta_alterdata: z.string().nullable().optional(),
    cdempresa_alterdata: z.string().nullable().optional(),
    show_alterdata: z.boolean().or(z.number().transform(val => Boolean(val))).nullable().optional(),
    serv_dorsal: z.string().nullable().optional(),
    bd_dorsal: z.string().nullable().optional(),
    login_dorsal: z.string().nullable().optional(),
    senha_dorsal: z.string().nullable().optional(),
    cdfilial: z.string().nullable().optional(),
    cdpdv: z.string().nullable().optional(),
    is_active: z.boolean().or(z.number().transform(val => Boolean(val))),
    is_system: z.boolean().or(z.number().transform(val => Boolean(val))).nullable().optional(),
    is_general_admin: z.boolean().or(z.number().transform(val => Boolean(val))).nullable().optional(),
    is_group_master: z.boolean().or(z.number().transform(val => Boolean(val))).nullable().optional(),
    company_group_id: z.number().int().positive().nullable().optional(),
    company_group_public_id: z.string().uuid().nullable().optional(),
    company_group_name: z.string().nullable().optional(),
    created_at: z.union([z.string(), z.date()]).optional(),
    updated_at: z.union([z.string(), z.date()]).optional(),
}).passthrough();